pyproject.toml
[project]
name = "ankismart"
version = "0.1.0"
description = "Intelligent Anki flashcard generation from documents"
requires-python = ">=3.11"
dependencies = [
    "pyside6>=6.6",
    "openai>=1.0",
    "python-docx>=1.0",
    "python-pptx>=0.6",
    "paddleocr>=2.7",
    "paddlepaddle>=2.5",
    "genanki>=0.13",
    "cryptography>=41.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "httpx>=0.25",
    "chardet>=5.0",
    "numpy>=1.26",
    "pillow>=10.0",
    "pypdfium2>=4.30",
]

[project.scripts]
ankismart = "ankismart.ui.app:main"

[dependency-groups]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.1",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["E501", "N802", "N806"]
